3139 Past project
this is a Past project that i did in my Final year. in it i was tasked in extending one of the three model presented there: Louis Voltkerra, SIR model, and coupled model for Guirrela warfare. i decide to extend the SIR model so that it account vaccination. Under the few assumption, i manipulated the vaccination rate to see the effect it has on death.